// packages/components/src/components/button-group.module.css.d.ts
declare const styles: {
  buttonGroup: string;
  "buttonGroup.horizontal": string;
  "buttonGroup.vertical": string;
  "buttonGroup.tight": string;
  "buttonGroup.normal": string;
  "buttonGroup.relaxed": string;
  "buttonGroup.ghost": string;
  buttonItem: string;
  "buttonItem.secondary": string;
  "buttonItem.outline": string;
  "buttonItem.ghost": string;
  "buttonItem.sm": string;
  "buttonItem.md": string;
  "buttonItem.lg": string;
  "buttonItem.icon": string;
  "buttonItem.icon.ghost": string;
  "buttonItem.selected": string;
  "buttonItem.secondary.selected": string;
  "buttonItem.outline.selected": string;
  "buttonItem.ghost.selected": string;
  itemWrapper: string;
  "itemWrapper.horizontal": string;
  "itemWrapper.vertical": string;
  "itemWrapper.first": string;
  "itemWrapper.last": string;
  "itemWrapper.first.horizontal": string;
  "itemWrapper.last.horizontal": string;
  "itemWrapper.first.vertical": string;
  "itemWrapper.last.vertical": string;
  horizontal: string;
  vertical: string;
  tight: string;
  normal: string;
  relaxed: string;
  ghost: string;
  secondary: string;
  outline: string;
  sm: string;
  md: string;
  lg: string;
  icon: string;
  selected: string;
  first: string;
  last: string;
};

export default styles;